d50

+++ includes/asm.h
@@ -184,7 +184,7 @@ void				set_cor_name(t_core *file, int fd);
 int					find_nbr_bytes(char *str, int size);
 void				set_cor_nbr_comment(t_core *file, int fd);
 int					result_opcode(char c, int code, int get_bin);
-int					negative_nbr(int nb, int size, t_core *file);
+long				negative_nbr(int nb, int size, t_core *file);
 void				set_instruction(t_core *file, t_cmd *c, int fd);
 void				set_bytes(int fd, char *str, t_core *file, t_cmd *c);
 void				set_bytes_zero_alignment(int fd, int size, int nb, int nbr);
